+++ manager/jam/src/main/java/org/yes/cart/service/endpoint/impl/PIMEndpointControllerImpl.java
@@ -15,30 +15,18 @@
  */
 package org.yes.cart.service.endpoint.impl;
 
-import org.apache.commons.lang.math.NumberUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.yes.cart.bulkjob.impl.BulkJobAutoContextImpl;
-import org.yes.cart.cluster.node.NodeService;
-import org.yes.cart.constants.AttributeNamesKeys;
 import org.yes.cart.domain.misc.MutablePair;
 import org.yes.cart.domain.vo.*;
-import org.yes.cart.service.async.impl.JobStatusListenerImpl;
-import org.yes.cart.service.async.model.AsyncContext;
-import org.yes.cart.service.async.model.JobContext;
-import org.yes.cart.service.async.model.JobContextKeys;
-import org.yes.cart.service.async.model.impl.JobContextImpl;
-import org.yes.cart.service.cluster.ClusterService;
+import org.yes.cart.service.cluster.ProductAsyncSupport;
 import org.yes.cart.service.endpoint.PIMEndpointController;
 import org.yes.cart.service.vo.VoProductService;
-import org.yes.cart.service.async.AsyncContextFactory;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 /**
  * User: denispavlov
@@ -49,19 +37,14 @@ import java.util.Map;
 public class PIMEndpointControllerImpl implements PIMEndpointController {
 
     private final VoProductService voProductService;
-    private final ClusterService clusterService;
-    private final AsyncContextFactory asyncContextFactory;
-    private final NodeService nodeService;
+    private final ProductAsyncSupport productAsyncSupport;
+
 
     @Autowired
     public PIMEndpointControllerImpl(final VoProductService voProductService,
-                                     final ClusterService clusterService,
-                                     final AsyncContextFactory asyncContextFactory,
-                                     final NodeService nodeService) {
+                                     final ProductAsyncSupport productAsyncSupport) {
         this.voProductService = voProductService;
-        this.clusterService = clusterService;
-        this.asyncContextFactory = asyncContextFactory;
-        this.nodeService = nodeService;
+        this.productAsyncSupport = productAsyncSupport;
     }
 
     @Override
@@ -86,7 +69,7 @@ public class PIMEndpointControllerImpl implements PIMEndpointController {
     public @ResponseBody
     VoProductWithLinks createProduct(@RequestBody final VoProduct vo) throws Exception {
         final VoProductWithLinks product = voProductService.createProduct(vo);
-        clusterService.reindexProduct(createJobContext(), product.getProductId());
+        productAsyncSupport.asyncIndexProduct(product.getProductId());
         return product;
     }
 
@@ -94,15 +77,18 @@ public class PIMEndpointControllerImpl implements PIMEndpointController {
     public @ResponseBody
     VoProductWithLinks updateProduct(@RequestBody final VoProductWithLinks vo) throws Exception {
         final VoProductWithLinks product = voProductService.updateProduct(vo);
-        clusterService.reindexProduct(createJobContext(), product.getProductId());
+        productAsyncSupport.asyncIndexProduct(product.getProductId());
         return product;
     }
 
     @Override
     public @ResponseBody
     void removeProduct(@PathVariable("id") final long id) throws Exception {
-        voProductService.removeProduct(id);
-        clusterService.reindexProduct(createJobContext(), id);
+        final VoProduct product = voProductService.getProductById(id);
+        if (product != null) {
+            voProductService.removeProduct(id);
+            productAsyncSupport.asyncIndexProduct(product.getProductId());
+        }
     }
 
     @Override
@@ -116,7 +102,7 @@ public class PIMEndpointControllerImpl implements PIMEndpointController {
     List<VoAttrValueProduct> updateProduct(@RequestBody final List<MutablePair<VoAttrValueProduct, Boolean>> vo) throws Exception {
         final List<VoAttrValueProduct> list = voProductService.updateProductAttributes(vo);
         if (!list.isEmpty()) {
-            clusterService.reindexProduct(createJobContext(), list.get(0).getProductId());
+            productAsyncSupport.asyncIndexProduct(list.get(0).getProductId());
         }
         return list;
     }
@@ -143,7 +129,7 @@ public class PIMEndpointControllerImpl implements PIMEndpointController {
     public @ResponseBody
     VoProductSku createSku(@RequestBody final VoProductSku vo) throws Exception {
         final VoProductSku sku = voProductService.createSku(vo);
-        clusterService.reindexProductSku(createJobContext(), sku.getSkuId());
+        productAsyncSupport.asyncIndexSku(sku.getSkuId());
         return sku;
     }
 
@@ -151,15 +137,18 @@ public class PIMEndpointControllerImpl implements PIMEndpointController {
     public @ResponseBody
     VoProductSku updateSku(@RequestBody final VoProductSku vo) throws Exception {
         final VoProductSku sku = voProductService.updateSku(vo);
-        clusterService.reindexProductSku(createJobContext(), sku.getSkuId());
+        productAsyncSupport.asyncIndexSku(sku.getSkuId());
         return sku;
     }
 
     @Override
     public @ResponseBody
     void removeSku(@PathVariable("id") final long id) throws Exception {
-        voProductService.removeSku(id);
-        clusterService.reindexProductSku(createJobContext(), id);
+        final VoProductSku sku = voProductService.getSkuById(id);
+        if (sku != null) {
+            voProductService.removeSku(id);
+            productAsyncSupport.asyncIndexSku(sku.getSkuId());
+        }
     }
 
     @Override
@@ -173,39 +162,9 @@ public class PIMEndpointControllerImpl implements PIMEndpointController {
     List<VoAttrValueProductSku> updateSku(@RequestBody final List<MutablePair<VoAttrValueProductSku, Boolean>> vo) throws Exception {
         final List<VoAttrValueProductSku> list = voProductService.updateSkuAttributes(vo);
         if (!list.isEmpty()) {
-            clusterService.reindexProductSku(createJobContext(), list.get(0).getSkuId());
+            productAsyncSupport.asyncIndexSku(list.get(0).getSkuId());
         }
         return list;
     }
-
-
-    private JobContext createJobContext() {
-
-        final Map<String, Object> param = new HashMap<>();
-        param.put(AsyncContext.TIMEOUT_KEY, AttributeNamesKeys.System.SYSTEM_CONNECTOR_PRODUCT_SINGLE_INDEX_TIMEOUT_MS);
-        param.put(JobContextKeys.NODE_FULL_PRODUCT_INDEX_STATE, new HashMap<String, Boolean>());
-        param.putAll(createAuthCtx().getAttributes());
-
-        // Max char of report to UI since it will get huge and simply will crash the UI, not to mention traffic cost.
-        final int logSize = NumberUtils.toInt(nodeService.getConfiguration().get(AttributeNamesKeys.System.IMPORT_JOB_LOG_SIZE), 100);
-        // Timeout - just in case runnable crashes and we need to unlock through timeout.
-        final int timeout = NumberUtils.toInt(nodeService.getConfiguration().get(AttributeNamesKeys.System.SYSTEM_CONNECTOR_PRODUCT_SINGLE_INDEX_TIMEOUT_MS), 100);
-
-        return new JobContextImpl(true, new JobStatusListenerImpl(logSize, timeout), param);
-    }
-
-    private AsyncContext createAuthCtx() {
-
-        final Map<String, Object> param = new HashMap<>();
-
-        try {
-            // This is manual access via Admin
-            return asyncContextFactory.getInstance(param);
-        } catch (IllegalStateException exp) {
-            // This is auto access with thread local
-            return new BulkJobAutoContextImpl(param);
-        }
-    }
-
 
 }
